requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
PyYAML>=6.0

# Opcional: Para etiquetado de imágenes
# labelme>=5.0.0

